chore: skip flaky SQL Lab test (#8282)
Tried 2-3 difference approaches to fix the flaky SQL Lab test that has
been biting us for a while. My guess is that the flakiness comes from
brace (the editor). Hoping a future release of brace and/or cypress
may help with this.

But for now, no test is better than a flaky one.