feat: support 'chevron' library for templating as jinja alternative
(chevron)[https://github.com/noahmorrison/chevron]
is a python implementation of mustache.js, and seems much safer
by scope than jinja could ever be.

While the library is not super active lately, it appears to be
feature complete and is easy to review from a security standpoint as
it's a few short-ish modules.

This PR adds support for the chevron library behind a feature flag.